refactor: remove `creator.rs`
